real fin translation
- 건물 목록 보러가기 버튼 번역
- 번역하기 전에 컨펌함

todo:
- 건물 이름들 다 번역하기 : 스프레드시트에 해두고 받아서 사용
- 예외처리. 검토해보고 예외 건물 몇개 없으면 그냥 처리, 많으면 json 만들어서 처리
- 편의시설 설명 번역

+++ mr-part-pj/src/components/NavBar.js
@@ -12,17 +12,23 @@ function NavBar() {
 
     const onChangeLang = () => {
         if (getLanguage() === "ko") {
-            window.localStorage.removeItem("i18nextLng")
-            window.localStorage.setItem("i18nextLng", 'en')
-            i18n.changeLanguage('en')
-            console.log('en')
+            if (window.confirm("Do you want to see the page in English?")) {
+                window.localStorage.removeItem("i18nextLng")
+                window.localStorage.removeItem("i18n")
+                window.localStorage.removeItem("i18nextlng")
+                window.localStorage.setItem("i18nextLng", 'en')
+                i18n.changeLanguage('en')
+            }
         }
         else {
-            window.localStorage.clear()
-            window.localStorage.removeItem("i18nextLng")
-            window.localStorage.setItem("i18nextLng", 'ko')
-            i18n.changeLanguage('ko')
-            console.log('ko')
+            if (window.confirm("한국어로 보시겠습니까?")){
+                window.localStorage.removeItem("i18nextLng")
+                window.localStorage.removeItem("i18n")
+                window.localStorage.removeItem("i18nextlng")
+                window.localStorage.setItem("i18nextLng", 'ko')
+                i18n.changeLanguage('ko')
+            }
+           
         }   
     }
